[step3/로또(자동)] refactor: StringUtil 패키지 변경
